refactor: move shared claim detail composables back to feature-claim-details
Since partner claims and regular claims share the same feature module,
there's no need to extract these components to ui-claim-status.
